backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.21.1
google-generativeai==0.7.2
pydantic==2.5.1
